# api/swagger.yaml
---
swagger: "2.0"
info: 
  version: "1.0.0"
  title: "Nintex shorturl application"

basePath: "/api"

tags:
- name: "Short URL application"
  description: |
    The API for creating and retrieving short url.
schemes:
- http

paths:
  /st/heartbeat:
    get:
      tags:
      - Operations
      operationId: heartbeat
      summary: "server heartbeat."
      description: "Simple health check."
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/heartbeat"
        500:
          description: "Unexpected Error"
          schema:
            $ref: "#/definitions/errors"
  
  /st:
    post:
      tags:
      - ShortUrl
      operationId: shorten
      description: "shortens url"
      parameters:
      - name: "x-request-id"
        in: "header"
        description: "correlation id, if not provided will be generated by the api"
        required: false
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/urlType"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/idType"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errors"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/errors"
        500:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/errors"
  
  /st/{id}:
    get:
      tags:
      - ShortUrl
      operationId: retrieve
      summary: "retrieving url using shorten url"
      description: "retrieving url using shorten url"
      parameters:
      - name: "x-request-id"
        in: "header"
        description: "correlation id, if not provided will be generated by the api"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "shortened id"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            $ref: "#/definitions/urlType"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errors"
        404:
          description: "NotFound"
          schema:
            $ref: "#/definitions/errors"
        500:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/errors"

definitions:
  heartbeat:
    type: "object"
    properties:
        message:
          type: "string"
  
  errors:
    description: Returns list of error(s) and their description.
    type: "object"
    properties:
      errors:
        type: "array"
        items:
          $ref: "#/definitions/errorModel"

  errorModel:
    type: "object"

  urlType:
    type: "object"
    properties:
      url:
        description: "URL for which short id needs to be generated"
        type: "string"
        pattern: https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)
    required:
    - url

  idType:
    type: "object"
    properties:
      id:
        description: "Unique short ID generated for the URL"
        type: "string"
    required:
    - id